+++ com.unity.render-pipelines.high-definition/Editor/Material/StackLit/ShaderGraph/StackLitGUI.cs
@@ -40,6 +40,8 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
         
         public static void SetupMaterialKeywordsAndPass(Material material)
         {
+            SynchronizeShaderGraphProperties(material);
+
             BaseLitGUI.SetupBaseLitKeywords(material);
             BaseLitGUI.SetupBaseLitMaterialPass(material);
             bool receiveSSR = material.HasProperty(kReceivesSSR) ? material.GetInt(kReceivesSSR) != 0 : false;
@@ -47,7 +49,6 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             BaseLitGUI.SetupStencil(material, receiveSSR, useSplitLighting);
         }
 
-        // Currently Lit material keyword setup is enough for StackLit so we don't have a function for it
         protected override void SetupMaterialKeywordsAndPassInternal(Material material) => SetupMaterialKeywordsAndPass(material);
     }
 }