+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed PBR master node preview
 - Now the static lighting sky will correctly take the default values for non-overridden properties
 - Fix shadergraph material pass setup not called
+- Fix property sync in shadergraph with the current material in the inspector
 
 ## [6.9.0] - 2019-07-02